+++ Bonobo.Git.Server/Controllers/GitController.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Diagnostics;
 using System.IO;
 using System.Net;
 using System.Web.Mvc;
@@ -11,6 +10,7 @@ using Bonobo.Git.Server.Models;
 using Bonobo.Git.Server.Security;
 using Ionic.Zlib;
 using Microsoft.Practices.Unity;
+using Serilog;
 using Repository = LibGit2Sharp.Repository;
 
 namespace Bonobo.Git.Server.Controllers
@@ -42,6 +42,7 @@ namespace Bonobo.Git.Server.Controllers
                 {
                     if (!RepositoryPermissionService.HasCreatePermission(User.Id()))
                     {
+                        Log.Warning("GitC: User {UserId} is not allowed to do push-to-create", User.Id());
                         return UnauthorizedResult();
                     }
                     if (!TryCreateOnPush(repositoryName))
@@ -62,6 +63,10 @@ namespace Bonobo.Git.Server.Controllers
             }
             else
             {
+                Log.Warning("GitC: SecureGetInfoRefs unauth because User {UserId} doesn't have permission {Permission} on  repo {RepositoryName}", 
+                    User.Id(),
+                    requiredLevel,
+                    repositoryName);
                 return UnauthorizedResult();
             }
         }
@@ -108,6 +113,7 @@ namespace Bonobo.Git.Server.Controllers
             if (directory.Exists)
             {
                 // We can't create a new repo - there's already a directory with that name
+                Log.Warning("GitC: Can't create {RepositoryName} - directory already exists", repositoryName);
                 return false;
             }
             RepositoryModel repository = new RepositoryModel();
@@ -115,6 +121,7 @@ namespace Bonobo.Git.Server.Controllers
             if (!repository.NameIsValid)
             {
                 // We don't like this name
+                Log.Warning("GitC: Can't create '{RepositoryName}' - name is invalid", repositoryName);
                 return false;
             }
             var user = MembershipService.GetUserModel(User.Id());
@@ -124,10 +131,12 @@ namespace Bonobo.Git.Server.Controllers
             if (!RepositoryRepository.Create(repository))
             {
                 // We can't add this to the repo store
+                Log.Warning("GitC: Can't create '{RepositoryName}' - RepoRepo.Create failed", repositoryName);
                 return false;
             }
 
             Repository.Init(Path.Combine(UserConfiguration.Current.Repositories, repository.Name), true);
+            Log.Information("GitC: '{RepositoryName}' created", repositoryName);
             return true;
         }
 
@@ -220,6 +229,10 @@ namespace Bonobo.Git.Server.Controllers
         {
             var directory = GetDirectoryInfo(repositoryName);
             var isValid = Repository.IsValid(directory.FullName);
+            if (!isValid)
+            {
+                Log.Warning("GitC: Invalid repo {RepositoryName}", repositoryName);
+            }
             return isValid;
         }
 
@@ -239,7 +252,7 @@ namespace Bonobo.Git.Server.Controllers
         protected override void OnException(ExceptionContext filterContext)
         {
             Exception exception = filterContext.Exception;
-            Trace.TraceError("{0}: Error caught in GitController {1}", DateTime.Now, exception);
+            Log.Error(exception, "Error caught in GitController");
             filterContext.Result = new ContentResult { Content = exception.ToString() };
 
             filterContext.ExceptionHandled = true;